#46: Move model classes into models package
