+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/ClaimDataExporter.java
@@ -1,5 +1,7 @@
 package ca.ualberta.cs.shinyexpensetracker;
 
+import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaim;
+
 // FIXME Changed from interface to abstract
 abstract public class ClaimDataExporter {
 	abstract public void export(ExpenseClaim claim);
