+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/WebServiceExporter.java
@@ -1,5 +1,7 @@
 package ca.ualberta.cs.shinyexpensetracker;
 
+import ca.ualberta.cs.shinyexpensetracker.models.ExpenseClaim;
+
 public class WebServiceExporter extends ClaimDataExporter {
 
 	@Override
