+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/models/ExpenseItem.java
@@ -1,7 +1,9 @@
-package ca.ualberta.cs.shinyexpensetracker;
+package ca.ualberta.cs.shinyexpensetracker.models;
 
 import java.util.ArrayList;
 import java.util.Date;
+
+import ca.ualberta.cs.shinyexpensetracker.IView;
 
 import android.graphics.Bitmap;
