+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/models/Tag.java
@@ -1,4 +1,4 @@
-package ca.ualberta.cs.shinyexpensetracker;
+package ca.ualberta.cs.shinyexpensetracker.models;
 
 public class Tag {
 	private String value;
